moved dealer buttons from Game to Table, added helper for isPlayer isDealer
