+++ src/lib/helper.ts
@@ -1,0 +1,9 @@
+import { NodeType } from "./constants"
+
+export const isDealer = (nodeType) => {
+    return nodeType === NodeType.dealer
+}
+
+export const isPlayer = (nodeType) => {
+    return nodeType === NodeType.player
+}
